+++ CHANGELOG.md
@@ -111,6 +111,7 @@
 - [8603](https://github.com/vegaprotocol/vega/issues/8603) - Restore `lastTradedPrice` of `nil` as `nil` in market snapshot
 - [8617](https://github.com/vegaprotocol/vega/issues/8617) - Fix panic with order gauge in future market
 - [8596](https://github.com/vegaprotocol/vega/issues/8596) - Fix panic when rejecting markets on time update.
+- [8545](https://github.com/vegaprotocol/vega/issues/6545) - Block explorer does not page backwards correctly.
 
 ## 0.71.0